+++ piwigoPublish.lrplugin/Resources/en/Strings.lua
@@ -1,0 +1,13 @@
+return {
+    piwigo_pluginTitle          = "Piwigo Publisher Plugin",
+    piwigo_pluginVersion        = "Plugin Version %s",
+    piwigo_selectService        = "Select publish service:",
+    piwigo_appliesToImages      = "Applies to selected images",
+    piwigo_setAlbumCover        = "Set Piwigo Album Cover",
+    piwigo_setAlbumCoverTooltip = "Sets selected image as Piwigo album cover",
+    piwigo_setAlbumCoverDesc    = "Sets selected image as Piwigo album cover for this collection",
+    piwigo_dialogTitle          = "Piwigo Extra Options",
+    piwigo_errorNoService       = "Could not find publish service",
+    piwigo_errorTitle           = "Error",
+    piwigo_closeButton          = "Close",
+}