+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+## [20251229.19] - 2025-12-29
+Fixed #22 - Error 1003 - Keyword already exisits. Required normalisation of keywords - all lower case, remove accents etc - for comparison between LrC and Piwigo keyword names as Piwigo effectively does accent-folding + case-folding and Lrc retains original.
+
+Fixed #16 (after re-opening) - New option to check local collection / set structure against existing Piwigo album structure - missing Piwigo albums will be created and incorrect links between collections / sets will be updated. Mis-named special collections will be renamed.
+
 ## [20251227.18] - 2025-12-27
 ### ### Fixed
 Fixed #16 - When importing a smart collection album a publishing error occurs - Running Import Smart Collection Settings created a collection within the Publish Service but a corresponding Piwigo album wasn't. This fix checks for missing albums during the Publish process and creates them if needed